+++ src/tools.cpp
@@ -1,8 +1,8 @@
 #include <iostream>
 #include "tools.h"
 
-using Eigen::VectorXd;
-using Eigen::MatrixXd;
+using Eigen::VectorXf;
+using Eigen::MatrixXf;
 using std::vector;
 
 
@@ -10,9 +10,9 @@ Tools::Tools() {}
 
 Tools::~Tools() {}
 
-VectorXd Tools::CalculateRMSE(const vector<VectorXd> &estimations,
-                              const vector<VectorXd> &ground_truth) {
-  VectorXd rmse(4);
+VectorXf Tools::CalculateRMSE(const vector<VectorXf> &estimations,
+                              const vector<VectorXf> &ground_truth) {
+  VectorXf rmse(4);
   rmse << 0, 0, 0, 0;
 
   unsigned long eSize = estimations.size();
@@ -22,7 +22,7 @@ VectorXd Tools::CalculateRMSE(const vector<VectorXd> &estimations,
   }
 
   for (unsigned int i = 0; i < eSize; i++) {
-    VectorXd residual = estimations[i] - ground_truth[i];
+    VectorXf residual = estimations[i] - ground_truth[i];
     residual = residual.array().pow(2);
     rmse += residual;
   }
@@ -33,16 +33,16 @@ VectorXd Tools::CalculateRMSE(const vector<VectorXd> &estimations,
   return rmse;
 }
 
-MatrixXd Tools::CalculateJacobian(const VectorXd &x_state) {
+MatrixXf Tools::CalculateJacobian(const VectorXf &x_state) {
   //Recover state parameters
-  float px = x_state(0);
-  float py = x_state(1);
-  float vx = x_state(2);
-  float vy = x_state(3);
+  float px = (float) x_state(0);
+  float py = (float) x_state(1);
+  float vx = (float) x_state(2);
+  float vy = (float) x_state(3);
 
-  float px_py = pow(px, 2) + pow(py, 2);
+  float px_py = (float) (pow(px, 2) + pow(py, 2));
 
-  MatrixXd Hj(3, 4);
+  MatrixXf Hj(3, 4);
 
   //Check division by zero
   if (px_py == 0) {
@@ -58,8 +58,8 @@ MatrixXd Tools::CalculateJacobian(const VectorXd &x_state) {
   }
 
   //Calculate our component terms
-  float root_px_py = sqrt(px_py);
-  float px_py_1_5 = pow(px_py, 1.5);
+  float root_px_py = (float) sqrt(px_py);
+  float px_py_1_5 = (float) pow(px_py, 1.5);
   float pyv = py * ((vx * py) - (vy * px));
   float pxv = px * ((vy * px) - (vx * py));
