+++ src/tools.h
@@ -18,12 +18,12 @@ public:
   /**
   * A helper method to calculate RMSE.
   */
-  Eigen::VectorXd CalculateRMSE(const std::vector<Eigen::VectorXd> &estimations, const std::vector<Eigen::VectorXd> &ground_truth);
+  Eigen::VectorXf CalculateRMSE(const std::vector<Eigen::VectorXf> &estimations, const std::vector<Eigen::VectorXf> &ground_truth);
 
   /**
   * A helper method to calculate Jacobians.
   */
-  Eigen::MatrixXd CalculateJacobian(const Eigen::VectorXd& x_state);
+  Eigen::MatrixXf CalculateJacobian(const Eigen::VectorXf& x_state);
 
 };
